新增：假日執行活動參與人員逐日可選上/下半天（半天以 0.5 天計）
參與執行人員的逐日參與日期，每個勾選日可再指定 全天／上午／下午。
chip 改為四態循環（未選 → 全天 → 上午 → 下午 → 未選），色系與版面沿用既有 pattern，
只有半天才在 chip 上疊白底小 pill 標示時段。申請人本人維持全程參與（不逐日、不半天）。

資料層
- TravelRequestParticipantDate 新增 Slot（full/am/pm，NOT NULL default 'full'）
  唯一索引維持 (ParticipantId, Date)，一天恆為一列
- TravelRequestParticipant.HolidayDays int? → decimal(5,1)?
- 權重單一真相 ParticipantDateSlots（full=1.0 / am=pm=0.5），null 與未知值一律視為全天
- Migration AddParticipantDateSlotAndDecimalHolidayDays（既有資料 3 → 3.0、Slot 補 'full'）

後端
- Create / Update 防禦分支 / Submit 三處的個人假日天數改為「假日 × 時段權重」加總
- ValidateParticipantDates 加 slot 值驗證與同日重複偵測（400，避免撞唯一索引變 500）
- FlexibleParticipantDateConverter：相容舊版前端送出的純日期字串（視為 full）
- Dapper 讀取 (int?) → (decimal?)，薪資 SQL UNION 兩分支顯式 CAST decimal(5,1)

金額
- 假日津貼改 MidpointRounding.AwayFromZero（奇數日薪 × .5 天必落中點，
  預設銀行家捨入會少 1 元）
- 一併修正簽核台既有落差：每人津貼預估原本一律套整單天數，
  改為 COALESCE(個人, 整單)，與薪資同一真相；HolidayAllowanceDto 加 Days 欄位

文件同步七處：CLAUDE.md / database-schema / application-forms / payroll-formula /
frontend-design（chip 四態循環變體）/ backend-design（payload 形狀演進 converter、
int → decimal 的 Dapper 執行期地雷）/ api-routes

+++ Api/Models/Dtos/PayrollDtos.cs
@@ -13,7 +13,7 @@ public sealed record EmployeePayrollDto(
     decimal   MealAllowance,
     decimal   OvertimePay,
     decimal   DailySalary,
-    int       HolidayTravelDays,
+    decimal   HolidayTravelDays,   // 半天以 0.5 計（假日活動參與人員可逐日勾上/下半天）
     decimal   HolidayAllowance,
     decimal   OtherAddition,
     string?   OtherAdditionNote,
